# service/jaxrs-server/src/gen/java/io/swagger/model/StorySummary.java
/*
 * StoryTime API
 * Backend for managing choose-your-own-destiny style stories.
 *
 * OpenAPI spec version: 0.3.0
 * Contact: saynotospam@ajegames.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.util.Date;
import javax.validation.constraints.*;

/**
 * StorySummary
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2017-05-12T22:56:19.776-07:00")
public class StorySummary   {
  @JsonProperty("storyKey")
  private String storyKey = null;

  @JsonProperty("authorId")
  private String authorId = null;

  @JsonProperty("version")
  private Integer version = null;

  @JsonProperty("title")
  private String title = null;

  @JsonProperty("penName")
  private String penName = null;

  @JsonProperty("tagLine")
  private String tagLine = null;

  @JsonProperty("about")
  private String about = null;

  @JsonProperty("firstChapter")
  private Integer firstChapter = null;

  @JsonProperty("publishedAt")
  private Date publishedAt = null;

  public StorySummary storyKey(String storyKey) {
    this.storyKey = storyKey;
    return this;
  }

   /**
   * Get storyKey
   * @return storyKey
  **/
  @JsonProperty("storyKey")
  @ApiModelProperty(value = "")
  public String getStoryKey() {
    return storyKey;
  }

  public void setStoryKey(String storyKey) {
    this.storyKey = storyKey;
  }

  public StorySummary authorId(String authorId) {
    this.authorId = authorId;
    return this;
  }

   /**
   * Get authorId
   * @return authorId
  **/
  @JsonProperty("authorId")
  @ApiModelProperty(value = "")
  public String getAuthorId() {
    return authorId;
  }

  public void setAuthorId(String authorId) {
    this.authorId = authorId;
  }

  public StorySummary version(Integer version) {
    this.version = version;
    return this;
  }

   /**
   * Get version
   * @return version
  **/
  @JsonProperty("version")
  @ApiModelProperty(value = "")
  public Integer getVersion() {
    return version;
  }

  public void setVersion(Integer version) {
    this.version = version;
  }

  public StorySummary title(String title) {
    this.title = title;
    return this;
  }

   /**
   * Get title
   * @return title
  **/
  @JsonProperty("title")
  @ApiModelProperty(value = "")
  public String getTitle() {
    return title;
  }

  public void setTitle(String title) {
    this.title = title;
  }

  public StorySummary penName(String penName) {
    this.penName = penName;
    return this;
  }

   /**
   * Get penName
   * @return penName
  **/
  @JsonProperty("penName")
  @ApiModelProperty(value = "")
  public String getPenName() {
    return penName;
  }

  public void setPenName(String penName) {
    this.penName = penName;
  }

  public StorySummary tagLine(String tagLine) {
    this.tagLine = tagLine;
    return this;
  }

   /**
   * Get tagLine
   * @return tagLine
  **/
  @JsonProperty("tagLine")
  @ApiModelProperty(value = "")
  public String getTagLine() {
    return tagLine;
  }

  public void setTagLine(String tagLine) {
    this.tagLine = tagLine;
  }

  public StorySummary about(String about) {
    this.about = about;
    return this;
  }

   /**
   * Get about
   * @return about
  **/
  @JsonProperty("about")
  @ApiModelProperty(value = "")
  public String getAbout() {
    return about;
  }

  public void setAbout(String about) {
    this.about = about;
  }

  public StorySummary firstChapter(Integer firstChapter) {
    this.firstChapter = firstChapter;
    return this;
  }

   /**
   * Get firstChapter
   * @return firstChapter
  **/
  @JsonProperty("firstChapter")
  @ApiModelProperty(value = "")
  public Integer getFirstChapter() {
    return firstChapter;
  }

  public void setFirstChapter(Integer firstChapter) {
    this.firstChapter = firstChapter;
  }

  public StorySummary publishedAt(Date publishedAt) {
    this.publishedAt = publishedAt;
    return this;
  }

   /**
   * Get publishedAt
   * @return publishedAt
  **/
  @JsonProperty("publishedAt")
  @ApiModelProperty(value = "")
  public Date getPublishedAt() {
    return publishedAt;
  }

  public void setPublishedAt(Date publishedAt) {
    this.publishedAt = publishedAt;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    StorySummary storySummary = (StorySummary) o;
    return Objects.equals(this.storyKey, storySummary.storyKey) &&
        Objects.equals(this.authorId, storySummary.authorId) &&
        Objects.equals(this.version, storySummary.version) &&
        Objects.equals(this.title, storySummary.title) &&
        Objects.equals(this.penName, storySummary.penName) &&
        Objects.equals(this.tagLine, storySummary.tagLine) &&
        Objects.equals(this.about, storySummary.about) &&
        Objects.equals(this.firstChapter, storySummary.firstChapter) &&
        Objects.equals(this.publishedAt, storySummary.publishedAt);
  }

  @Override
  public int hashCode() {
    return Objects.hash(storyKey, authorId, version, title, penName, tagLine, about, firstChapter, publishedAt);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class StorySummary {\n");
    
    sb.append("    storyKey: ").append(toIndentedString(storyKey)).append("\n");
    sb.append("    authorId: ").append(toIndentedString(authorId)).append("\n");
    sb.append("    version: ").append(toIndentedString(version)).append("\n");
    sb.append("    title: ").append(toIndentedString(title)).append("\n");
    sb.append("    penName: ").append(toIndentedString(penName)).append("\n");
    sb.append("    tagLine: ").append(toIndentedString(tagLine)).append("\n");
    sb.append("    about: ").append(toIndentedString(about)).append("\n");
    sb.append("    firstChapter: ").append(toIndentedString(firstChapter)).append("\n");
    sb.append("    publishedAt: ").append(toIndentedString(publishedAt)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}